requirements.txt
pydantic 
langchain
langchain-groq
langgraph
streamlit
python-dotenv
langchain-core
langchain-community
langchain-ollama

